requirements.txt
fastapi[all]
uvicorn
python-multipart
Pillow